refine: refine create dataset code

+++ create_dataset.py
@@ -1,16 +1,13 @@
 from dataclasses import dataclass
 import os
-import pickle
-import random
 import sys
-from typing import Sequence, cast
+from typing import cast
 
 import numpy as np
 
 
 from retina.face import FACIAL_EXPRESSION_MAPPER, FacialExpressionLabel
 from retina.log import Ansi
-from retina.size import Dimension
 
 sys.path.append(os.getcwd())
 
@@ -21,20 +18,14 @@ import tqdm
 import cv2 as cv
 import pandas as pd
 import numpy.typing as npt
-import sklearn.decomposition
 
 import retina
 
-if (os.path.exists(retina.filesys.DATA_DIR_PATH)):
-  shutil.rmtree(retina.filesys.DATA_DIR_PATH)
-with zipfile.ZipFile(retina.filesys.DATASET_PATH, 'r') as zip_ref:
-  zip_ref.extractall(retina.filesys.DATA_DIR_PATH)
-
-def preprocess_image(img: cv.typing.MatLike):
-  img = retina.cvutil.resize_image(original, retina.size.STANDARD_DIMENSIONS) # Resize
-  img = cv.cvtColor(img, cv.COLOR_BGR2GRAY) # Grayscale
-  img = retina.colors.clahe(img) # Contrast adjustment
-  return img
+if "--unzip" in sys.argv:
+  if (os.path.exists(retina.filesys.DATA_DIR_PATH)):
+    shutil.rmtree(retina.filesys.DATA_DIR_PATH)
+  with zipfile.ZipFile(retina.filesys.DATASET_PATH, 'r') as zip_ref:
+    zip_ref.extractall(retina.filesys.DATA_DIR_PATH)
 
 @dataclass
 class TrainDataEntry:
@@ -55,44 +46,34 @@ for folder in os.scandir(retina.filesys.DATA_DIR_PATH):
     retina.filesys.get_files_in_folder(folder.path)
   ))
 
+skipped = 0
 list_data: list[npt.NDArray] = []
 list_labels: list[int] = []
 for entry in tqdm.tqdm(entries, desc="Building dataset from images"):
   original = cv.imread(entry.path)
-  img = preprocess_image(original)
-  canvas = img.copy()
-  faces, _ = retina.face.extract_faces(img)
-
-  if len(faces) == 0:
+  canvas = cv.cvtColor(retina.face.preprocess_face_image(original), cv.COLOR_GRAY2BGR)
+  features = retina.face.face2vec(original, canvas=canvas)
+  retina.debug.imdebug(canvas)
+  if features is None:
+    print(f"\n{Ansi.Error}Skipping {entry.path} because no faces were found in the image.{Ansi.End}")
+    skipped += 1
     continue
-
-  for face in faces:
-    face = cv.resize(face, retina.size.FACE_DIMENSIONS.tuple, interpolation=cv.INTER_CUBIC)
-    canvas = cv.cvtColor(face, cv.COLOR_GRAY2BGR)
-    landmark = retina.face.extract_face_landmarks(face, canvas=canvas)
-    face = retina.face.face_alignment(face, landmark)
-    texture = retina.face.grid_lbp(face)
-    retina.debug.imdebug(canvas)
-
-    list_data.append(texture)
+  for feature in features:
+    list_data.append(feature)
     list_labels.append(entry.label.value)
   
 if len(list_data) == 0:
   print(f"{Ansi.Error}No images were successfully processed into the dataset.{Ansi.End}")
   exit(1)
 
+print(f"{Ansi.Warning}Skipped over {skipped} images because no faces were found in the images.")
+
 data = np.array(list_data)
-# Normalization
-data = data / data.sum(axis=1).reshape((-1, 1))
-
 labels = np.array(list_labels).reshape((-1, 1))
-
 dfdata = np.hstack((labels, data))
-
-
 df = pd.DataFrame(dfdata, columns=[
   "label",
-  *map(lambda idx: f'texture-{idx + 1}', range(data.shape[1])),
+  *map(lambda idx: f'feature-{idx + 1}', range(data.shape[1])),
 ])
 
 df.to_csv(retina.filesys.TRAINING_DATA_CSV_PATH, index=False)
